:muscle: Merge `is/*.ts` into `is.ts`
Owing to the principle outlined below, defining everything within the `is.ts`
file is considerably more straightforward.
https://jsr.io/docs/about-slow-types#types-must-be-simply-inferred-or-explicit
