--- is/annotation_test.ts
@@ -1,340 +1,0 @@
-import {
-  assertEquals,
-  assertStrictEquals,
-} from "https://deno.land/std@0.211.0/assert/mod.ts";
-import {
-  assertSnapshot,
-} from "https://deno.land/std@0.211.0/testing/snapshot.ts";
-import { assertType } from "https://deno.land/std@0.211.0/testing/types.ts";
-import { type Equal, stringify } from "./_testutil.ts";
-import type { Predicate } from "./type.ts";
-import {
-  isArray,
-  isAsyncFunction,
-  isBigInt,
-  isBoolean,
-  isFunction,
-  isNull,
-  isNumber,
-  isRecordObject,
-  isSet,
-  isString,
-  isSymbol,
-  isSyncFunction,
-  isUndefined,
-} from "./core.ts";
-import { isObjectOf, isTupleOf, isUniformTupleOf } from "./factory.ts";
-import is, {
-  isOptionalOf,
-  isReadonlyOf,
-  isUnwrapOptionalOf,
-  isUnwrapReadonlyOf,
-} from "./annotation.ts";
-
-const examples = {
-  string: ["", "Hello world"],
-  number: [0, 1234567890],
-  bigint: [0n, 1234567890n],
-  boolean: [true, false],
-  array: [[], [0, 1, 2], ["a", "b", "c"], [0, "a", true]],
-  set: [new Set(), new Set([0, 1, 2]), new Set(["a", "b", "c"])],
-  record: [{}, { a: 0, b: 1, c: 2 }, { a: "a", b: "b", c: "c" }],
-  map: [
-    new Map(),
-    new Map([["a", 0], ["b", 1], ["c", 2]]),
-    new Map([["a", "a"], ["b", "b"], ["c", "c"]]),
-  ],
-  syncFunction: [function a() {}, () => {}],
-  asyncFunction: [async function b() {}, async () => {}],
-  null: [null],
-  undefined: [undefined],
-  symbol: [Symbol("a"), Symbol("b"), Symbol("c")],
-  date: [new Date(1690248225000), new Date(0)],
-  promise: [new Promise(() => {})],
-} as const;
-
-async function testWithExamples<T>(
-  t: Deno.TestContext,
-  pred: Predicate<T>,
-  opts?: {
-    validExamples?: (keyof typeof examples)[];
-    excludeExamples?: (keyof typeof examples)[];
-  },
-): Promise<void> {
-  const { validExamples = [], excludeExamples = [] } = opts ?? {};
-  const exampleEntries = (Object.entries(examples) as unknown as [
-    name: keyof typeof examples,
-    example: unknown[],
-  ][]).filter(([k]) => !excludeExamples.includes(k));
-  for (const [name, example] of exampleEntries) {
-    const expect = validExamples.includes(name);
-    for (const v of example) {
-      await t.step(
-        `returns ${expect} on ${stringify(v)}`,
-        () => {
-          assertEquals(pred(v), expect);
-        },
-      );
-    }
-  }
-}
-
-Deno.test("isOptionalOf<T>", async (t) => {
-  await t.step("returns properly named function", async (t) => {
-    await assertSnapshot(t, isOptionalOf(isNumber).name);
-    // Nesting does nothing
-    await assertSnapshot(t, isOptionalOf(isOptionalOf(isNumber)).name);
-  });
-  await t.step("returns proper type predicate", () => {
-    const a: unknown = undefined;
-    if (isOptionalOf(isNumber)(a)) {
-      assertType<Equal<typeof a, number | undefined>>(true);
-    }
-  });
-  await t.step("with isString", async (t) => {
-    await testWithExamples(t, isOptionalOf(isString), {
-      validExamples: ["string", "undefined"],
-    });
-  });
-  await t.step("with isNumber", async (t) => {
-    await testWithExamples(t, isOptionalOf(isNumber), {
-      validExamples: ["number", "undefined"],
-    });
-  });
-  await t.step("with isBigInt", async (t) => {
-    await testWithExamples(t, isOptionalOf(isBigInt), {
-      validExamples: ["bigint", "undefined"],
-    });
-  });
-  await t.step("with isBoolean", async (t) => {
-    await testWithExamples(t, isOptionalOf(isBoolean), {
-      validExamples: ["boolean", "undefined"],
-    });
-  });
-  await t.step("with isArray", async (t) => {
-    await testWithExamples(t, isOptionalOf(isArray), {
-      validExamples: ["array", "undefined"],
-    });
-  });
-  await t.step("with isSet", async (t) => {
-    await testWithExamples(t, isOptionalOf(isSet), {
-      validExamples: ["set", "undefined"],
-    });
-  });
-  await t.step("with isRecordObject", async (t) => {
-    await testWithExamples(t, isOptionalOf(isRecordObject), {
-      validExamples: ["record", "undefined"],
-    });
-  });
-  await t.step("with isFunction", async (t) => {
-    await testWithExamples(t, isOptionalOf(isFunction), {
-      validExamples: ["syncFunction", "asyncFunction", "undefined"],
-    });
-  });
-  await t.step("with isSyncFunction", async (t) => {
-    await testWithExamples(t, isOptionalOf(isSyncFunction), {
-      validExamples: ["syncFunction", "undefined"],
-    });
-  });
-  await t.step("with isAsyncFunction", async (t) => {
-    await testWithExamples(t, isOptionalOf(isAsyncFunction), {
-      validExamples: ["asyncFunction", "undefined"],
-    });
-  });
-  await t.step("with isNull", async (t) => {
-    await testWithExamples(t, isOptionalOf(isNull), {
-      validExamples: ["null", "undefined"],
-    });
-  });
-  await t.step("with isUndefined", async (t) => {
-    await testWithExamples(t, isOptionalOf(isUndefined), {
-      validExamples: ["undefined"],
-    });
-  });
-  await t.step("with isSymbol", async (t) => {
-    await testWithExamples(t, isOptionalOf(isSymbol), {
-      validExamples: ["symbol", "undefined"],
-    });
-  });
-});
-
-Deno.test("isUnwrapOptionalOf<T>", async (t) => {
-  await t.step("returns properly named function", async (t) => {
-    await assertSnapshot(t, isUnwrapOptionalOf(isOptionalOf(isNumber)).name);
-    // Non optional does nothing
-    await assertSnapshot(t, isUnwrapOptionalOf(isNumber).name);
-    // Nesting does nothing
-    await assertSnapshot(
-      t,
-      isUnwrapOptionalOf(isUnwrapOptionalOf(isOptionalOf(isNumber))).name,
-    );
-  });
-  await t.step("returns proper type predicate", () => {
-    const a: unknown = undefined;
-    if (isUnwrapOptionalOf(isOptionalOf(isNumber))(a)) {
-      assertType<Equal<typeof a, number>>(true);
-    }
-    if (isUnwrapOptionalOf(isNumber)(a)) {
-      assertType<Equal<typeof a, number>>(true);
-    }
-  });
-  await t.step("with isString", async (t) => {
-    await testWithExamples(t, isUnwrapOptionalOf(isOptionalOf(isString)), {
-      validExamples: ["string"],
-    });
-  });
-  await t.step("with isNumber", async (t) => {
-    await testWithExamples(t, isUnwrapOptionalOf(isOptionalOf(isNumber)), {
-      validExamples: ["number"],
-    });
-  });
-  await t.step("with isBigInt", async (t) => {
-    await testWithExamples(t, isUnwrapOptionalOf(isOptionalOf(isBigInt)), {
-      validExamples: ["bigint"],
-    });
-  });
-  await t.step("with isBoolean", async (t) => {
-    await testWithExamples(t, isUnwrapOptionalOf(isOptionalOf(isBoolean)), {
-      validExamples: ["boolean"],
-    });
-  });
-  await t.step("with isArray", async (t) => {
-    await testWithExamples(t, isUnwrapOptionalOf(isOptionalOf(isArray)), {
-      validExamples: ["array"],
-    });
-  });
-  await t.step("with isSet", async (t) => {
-    await testWithExamples(t, isUnwrapOptionalOf(isOptionalOf(isSet)), {
-      validExamples: ["set"],
-    });
-  });
-  await t.step("with isRecordObject", async (t) => {
-    await testWithExamples(
-      t,
-      isUnwrapOptionalOf(isOptionalOf(isRecordObject)),
-      {
-        validExamples: ["record"],
-      },
-    );
-  });
-  await t.step("with isFunction", async (t) => {
-    await testWithExamples(t, isUnwrapOptionalOf(isOptionalOf(isFunction)), {
-      validExamples: ["syncFunction", "asyncFunction"],
-    });
-  });
-  await t.step("with isSyncFunction", async (t) => {
-    await testWithExamples(
-      t,
-      isUnwrapOptionalOf(isOptionalOf(isSyncFunction)),
-      {
-        validExamples: ["syncFunction"],
-      },
-    );
-  });
-  await t.step("with isAsyncFunction", async (t) => {
-    await testWithExamples(
-      t,
-      isUnwrapOptionalOf(isOptionalOf(isAsyncFunction)),
-      {
-        validExamples: ["asyncFunction"],
-      },
-    );
-  });
-  await t.step("with isNull", async (t) => {
-    await testWithExamples(t, isUnwrapOptionalOf(isOptionalOf(isNull)), {
-      validExamples: ["null"],
-    });
-  });
-  await t.step("with isUndefined", async (t) => {
-    await testWithExamples(t, isUnwrapOptionalOf(isOptionalOf(isUndefined)), {
-      validExamples: ["undefined"],
-    });
-  });
-  await t.step("with isSymbol", async (t) => {
-    await testWithExamples(t, isUnwrapOptionalOf(isOptionalOf(isSymbol)), {
-      validExamples: ["symbol"],
-    });
-  });
-});
-
-Deno.test("isReadonlyOf<T>", async (t) => {
-  await t.step("returns properly named function", async (t) => {
-    await assertSnapshot(t, isReadonlyOf(isNumber).name);
-    // Nesting does nothing
-    await assertSnapshot(t, isReadonlyOf(isReadonlyOf(isNumber)).name);
-  });
-  await t.step("returns proper type predicate", () => {
-    const a: unknown = undefined;
-    if (isReadonlyOf(isNumber)(a)) {
-      assertType<Equal<typeof a, Readonly<number>>>(true);
-    }
-    if (isReadonlyOf(isTupleOf([isString, isNumber, isBoolean]))(a)) {
-      assertType<Equal<typeof a, Readonly<[string, number, boolean]>>>(true);
-    }
-    if (isReadonlyOf(isUniformTupleOf(3, isString))(a)) {
-      assertType<Equal<typeof a, Readonly<[string, string, string]>>>(true);
-    }
-    if (
-      isReadonlyOf(isObjectOf({ a: isString, b: isNumber, c: isBoolean }))(a)
-    ) {
-      assertType<
-        Equal<typeof a, Readonly<{ a: string; b: number; c: boolean }>>
-      >(true);
-    }
-  });
-});
-
-Deno.test("isUnwrapReadonlyOf<T>", async (t) => {
-  await t.step("returns properly named function", async (t) => {
-    await assertSnapshot(t, isUnwrapReadonlyOf(isReadonlyOf(isNumber)).name);
-    // Nesting does nothing
-    await assertSnapshot(
-      t,
-      isUnwrapReadonlyOf(isReadonlyOf(isReadonlyOf(isNumber))).name,
-    );
-  });
-  await t.step("returns proper type predicate", () => {
-    const a: unknown = undefined;
-    if (isUnwrapReadonlyOf(isReadonlyOf(isNumber))(a)) {
-      assertType<Equal<typeof a, number>>(true);
-    }
-    if (
-      isUnwrapReadonlyOf(
-        isReadonlyOf(isTupleOf([isString, isNumber, isBoolean])),
-      )(a)
-    ) {
-      assertType<Equal<typeof a, [string, number, boolean]>>(true);
-    }
-    if (isUnwrapReadonlyOf(isReadonlyOf(isUniformTupleOf(3, isString)))(a)) {
-      assertType<Equal<typeof a, [string, string, string]>>(true);
-    }
-    if (
-      isUnwrapReadonlyOf(
-        isReadonlyOf(isObjectOf({ a: isString, b: isNumber, c: isBoolean })),
-      )(a)
-    ) {
-      assertType<
-        Equal<typeof a, { a: string; b: number; c: boolean }>
-      >(true);
-    }
-  });
-});
-
-Deno.test("is", async (t) => {
-  const mod = await import("./annotation.ts");
-  const casesOfAliasAndIsFunction = Object.entries(mod)
-    .filter(([k, _]) => k.startsWith("is"))
-    .map(([k, v]) => [k.slice(2), v] as const);
-  for (const [alias, fn] of casesOfAliasAndIsFunction) {
-    await t.step(`defines \`${alias}\` function`, () => {
-      assertStrictEquals(is[alias as keyof typeof is], fn);
-    });
-  }
-  await t.step(
-    "only has entries that are the same as the `is*` function aliases",
-    () => {
-      const aliases = casesOfAliasAndIsFunction.map(([a]) => a).sort();
-      assertEquals(Object.keys(is).sort(), aliases);
-    },
-  );
-});
